requirements.txt
streamlit>=1.22.0
pandas>=1.3.0
numpy>=1.21.0
plotly>=5.0.0
python-dotenv>=0.19.0 
mysql-connector-python>=8.0.0
pyodbc>=4.0.0
matplotlib>=3.0.0
python-dateutil>=2.8.0 
